backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
redis